refactor(OMN-14289): retire validation→nodes import-layering back-edge in omnibase_core
Burn down the frozen `validation -> nodes` back-edge in the `.importlinter`
oracle (`core-midtier-no-top` contract, OMN-14216). validator_routing_authority.py
imported ModelResidueEntry, ModelRoutingAuthorityCheckOutput, and
check_routing_authority_at_path from the nodes package.

Real dependency inversion, not a suppression:
- Extract the pure check algorithm (all _scan_*/_build_* helpers, the
  positive/negative/residue/shape audits, and the check_routing_authority_at_path
  I/O helper) out of nodes/node_routing_authority_check_compute/handler.py into
  a new validation-layer module: validation/checker_routing_authority.py.
- NodeRoutingAuthorityCheckCompute becomes a thin wrapper delegating to that
  module (nodes -> validation is a doctrine-legal direction).
- validator_routing_authority.py now imports the two models directly from
  their canonical models.validation.* locations and check_routing_authority_at_path
  from the new validation module — zero imports from omnibase_core.nodes remain.
- Drop the now-unmatched validation -> nodes ignore line from .importlinter
  (unmatched_ignore_imports_alerting=ERROR ratchet).

Public API of NodeRoutingAuthorityCheckCompute (.check(), ._resolve_endpoint_for_ref())
is unchanged, so existing tests pass without modification.

Refs OMN-3210
